FatModelの整備
