+++ web/app/Models/CheckedTodo.php
@@ -7,6 +7,8 @@ use Illuminate\Database\Eloquent\Model;
 use LINE\LINEBot\MessageBuilder\TemplateBuilder\ButtonTemplateBuilder;
 use LINE\LINEBot\TemplateActionBuilder\PostbackTemplateActionBuilder;
 use LINE\LINEBot\MessageBuilder\TemplateMessageBuilder;
+use LINE\LINEBot\MessageBuilder\TemplateBuilder\ConfirmTemplateBuilder;
+use LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
 
 class CheckedTodo extends Model
 {
@@ -58,5 +60,117 @@ class CheckedTodo extends Model
                 )
             );
         return $builder;
+    }
+
+    /**
+     * CheckTodoのカルーセル
+     *
+     * @param object $todo
+     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
+     */
+    public static function createCheckTodoCarouselColumn(object $todo)
+    {
+        if ($todo->depth === "0") {
+            $parent = 'プロジェクト:「' . $todo->project->name . '」のゴール';
+        } else {
+            $parentTodo = Todo::where('uuid', $todo->parent_uuid)->first();
+            $parent = '「' . $parentTodo->name . '」のためにやること';
+        }
+        $accomplish = count($todo->accomplish) > 0  ? '【達成】' : '【未達成】';
+        $title = $accomplish . $todo->name;
+        $actions = [
+            new PostbackTemplateActionBuilder('振り返る', 'action=CHECK_TODO&todo_uuid=' . $todo->uuid),
+        ];
+        $builder = new CarouselColumnTemplateBuilder($title, $parent, null, $actions);
+        return $builder;
+    }
+
+    /**
+     * Todoが達成したかどうか
+     *
+     * @param Todo $todo
+     * @return \LINE\LINEBot\MessageBuilder\TemplateBuilder\ConfirmTemplateBuilder
+     */
+    public static function askIfTodoHasBeenAccomplished(Todo $todo)
+    {
+        $text = '「' . $todo->name . '」について達成できましたか？';
+        $builder =
+            new ConfirmTemplateBuilder(
+                $text,
+                [
+                    new PostbackTemplateActionBuilder('はい', 'action=ACOMMPLISHED_TODO&todo_uuid=' . $todo->uuid),
+                    new PostbackTemplateActionBuilder('いいえ', 'action=NOT_ACCOMPLISHED_TODO&todo_uuid=' . $todo->uuid)
+                ]
+            );
+        return $builder;
+    }
+
+    /**
+     * Todoが達成しなかった時、そのTodoを達成させるために新しくTodoを追加するかどうか
+     *
+     * @param Todo $todo
+     * @return \LINE\LINEBot\MessageBuilder\TemplateBuilder\ConfirmTemplateBuilder
+     */
+    public static function addTodoAfterCheckTodo(Todo $todo)
+    {
+        $text = '「' . $todo->name . '」を達成するためにやることを新しく追加しますか?';
+        $builder =
+            new ConfirmTemplateBuilder(
+                $text,
+                [
+                    new PostbackTemplateActionBuilder('はい', 'action=ADD_TODO_AFTER_CHECK_TODO&todo_uuid=' . $todo->uuid),
+                    new PostbackTemplateActionBuilder('いいえ', 'action=NOT_ADD_TODO_AFTER_CHECK_TODO&todo_uuid=' . $todo->uuid)
+                ]
+            );
+        return $builder;
+    }
+
+    /**
+     * 振り返りを続けるかどうか
+     *
+     * @param LineUsersQuestion $question
+     * @return \LINE\LINEBot\MessageBuilder\TemplateBuilder\ConfirmTemplateBuilder
+     */
+    public static function askContinueCheckTodo(LineUsersQuestion $question)
+    {
+        if ($question->checked_todo === CheckedTodo::CHECK_TODO['CHECK_TODO_BY_TODAY']) {
+            $title = '今日までにやることの振り返りを続けますか？';
+            $action_type = 'CHECK_TODO_BY_TODAY';
+        } else if ($question->checked_todo === CheckedTodo::CHECK_TODO['CHECK_TODO_BY_THIS_WEEK']) {
+            $title = '今週までにやることの振り返りを続けますか？';
+            $action_type = 'CHECK_TODO_BY_THIS_WEEK';
+        } else if ($question->checked_todo === CheckedTodo::CHECK_TODO['SELECT_TODO_LIST_TO_CHECK']) {
+            $title = '振り返りを続けますか？';
+            $action_type = 'SELECT_TODO_LIST_TO_CHECK';
+        }
+        $builder =
+            new ConfirmTemplateBuilder(
+                $title, // text
+                [
+                    new PostbackTemplateActionBuilder('続ける', 'action=' . $action_type . '&todo_uuid='),
+                    new PostbackTemplateActionBuilder('終了する', 'action=FINISH_CHECK_TODO&todo_uuid='),
+                ]
+            );
+        return $builder;
+    }
+
+
+
+    /**
+     * 振り返り終了のアナウンス
+     *
+     * @param LineUsersQuestion $question
+     * @return string $text
+     */
+    public static function getTextMessageOfFinishCheckTodo(LineUsersQuestion $question)
+    {
+        if ($question->checked_todo === CheckedTodo::CHECK_TODO['CHECK_TODO_BY_TODAY']) {
+            $text = '今日までにやることの振り返りを終了しました。';
+        } else if ($question->checked_todo === CheckedTodo::CHECK_TODO['CHECK_TODO_BY_THIS_WEEK']) {
+            $text = '今週までにやることの振り返りを終了しました。';
+        } else if ($question->checked_todo === CheckedTodo::CHECK_TODO['SELECT_TODO_LIST_TO_CHECK']) {
+            $text = '振り返りを終了しました。';
+        }
+        return $text;
     }
 }
