+++ web/app/Models/Todo.php
@@ -64,6 +64,13 @@ class Todo extends Model
         return $this->hasMany(CheckedTodo::class, 'todo_uuid', 'uuid');
     }
 
+
+    /**
+     *
+     * ゴール Goal
+     *
+     */
+
     /**
      * プロジェクトのゴールを聞く
      *
@@ -76,6 +83,15 @@ class Todo extends Model
         return '今度は' . $user_name . 'さんが「' . $project_name . '」で達成したいゴールを教えて！';
     }
 
+
+    /**
+     *
+     *
+     * やること Todo
+     *
+     *
+     */
+
     /**
      * Todoの名前を聞く
      *
@@ -87,6 +103,118 @@ class Todo extends Model
         return '「' . $todo->name . '」を達成するためにやることを教えてください！';
     }
 
+    /**
+     * 変更後のTodoの名前を聞く
+     *
+     * @param Todo $todo
+     * @return string $reply_message
+     */
+    public static function askTodoReName(Todo $todo)
+    {
+        return 'やること:「' . $todo->name . '」の変更後の名前を教えてください！';
+    }
+
+    /**
+     * 名前の変更を確認する
+     *
+     * @param Todo $todo
+     * @param DateTime $date
+     * @return string $reply_message
+     */
+    public static function reportNewTodoName(Todo $todo, string $new_todo_name)
+    {
+        return '「' . $todo->name . '」から「' . $new_todo_name . '」へ変更が完了しました';
+    }
+
+    /**
+     * Todoのカルーセル
+     *
+     * @param object $todo
+     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
+     */
+    public static function createTodoCarouselColumn(object $todo)
+    {
+        if ($todo->depth === "0") {
+            $parent = 'プロジェクト:「' . $todo->project->name . '」のゴール';
+        } else {
+            $parentTodo = Todo::where('uuid', $todo->parent_uuid)->first();
+            $parent = '「' . $parentTodo->name . '」のためにやること';
+        }
+        $accomplish = count($todo->accomplish) > 0 ? '【達成】' : '【未達成】';
+        $title = $accomplish . $todo->name;
+        $actions = [
+            new PostbackTemplateActionBuilder('名前・期限の変更/削除', 'action=CHANGE_TODO&todo_uuid=' . $todo->uuid),
+            new PostbackTemplateActionBuilder('やることの追加', 'action=ADD_TODO&todo_uuid=' . $todo->uuid),
+            new PostbackTemplateActionBuilder('振り返る', 'action=SELECT_CHECK_TODO&todo_uuid=' . $todo->uuid),
+        ];
+        $builder = new CarouselColumnTemplateBuilder($title, $parent, null, $actions);
+        return $builder;
+    }
+
+    /**
+     * 作ったTodoのTodoを新しく追加する
+     *
+     * @param object $todo
+     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
+     */
+    public static function continueAddTodoOfTodo(object $todo)
+    {
+        $carouselText =  '「' . $todo->name . '」を達成するためにやることを新しく追加しますか?';
+        $actions = [
+            new PostbackTemplateActionBuilder('追加する', 'action=ADD_TODO&todo_uuid=' . $todo->uuid),
+        ];
+        $builder = new CarouselColumnTemplateBuilder(null, $carouselText, null, $actions);
+        return $builder;
+    }
+
+    /**
+     * 引き続き親TodoのTodoを追加する
+     *
+     * @param object $parentTodo
+     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
+     */
+    public static function continueAddTodoOfParentTodo(object $parent_todo)
+    {
+        $carouselText =  '「' . $parent_todo->name . '」を達成するためにやることを引き続き追加しますか?';
+        $actions = [
+            new PostbackTemplateActionBuilder('追加する', 'action=ADD_TODO&todo_uuid=' . $parent_todo->uuid),
+        ];
+        $builder = new CarouselColumnTemplateBuilder(null, $carouselText, null, $actions);
+        return $builder;
+    }
+
+
+    /**
+     * Todoを変更する
+     *
+     * @param Todo $todo
+     * @return \LINE\LINEBot\MessageBuilder\TemplateMessageBuilder
+     */
+    public static function changeTodo(Todo $todo)
+    {
+        $title =  '「' . $todo->name . '」';
+        return new TemplateMessageBuilder(
+            $title,
+            new ButtonTemplateBuilder(
+                $title,
+                "選択してください",
+                null,
+                [
+                    new PostbackTemplateActionBuilder("名前の変更", 'action=RENAME_TODO&todo_uuid=' . $todo->uuid),
+                    new PostbackTemplateActionBuilder('やることの削除', 'action=DELETE_TODO&todo_uuid=' . $todo->uuid),
+                    new PostbackTemplateActionBuilder('期限の変更', 'action=ASK_RESCHEDULE&todo_uuid=' . $todo->uuid),
+                ]
+            )
+        );
+    }
+
+    /**
+     *
+     *
+     * やること一覧 TodoList
+     *
+     *
+     */
 
     /**
      * Todoの一覧を見るか、それとも新しくTodoを追加するか尋ねる
@@ -112,6 +240,60 @@ class Todo extends Model
             );
         return $builder;
     }
+
+    /**
+     * Todoの一覧表示
+     *
+     * @param User $line_user
+     * @param string $action_value
+     * @param array $todo_carousel_columns
+     * @return \LINE\LINEBot\MessageBuilder\TemplateMessageBuilder
+     */
+    public static function createTodoListTitleMessage(User $line_user, string $action_value, array $todo_carousel_columns)
+    {
+        if (
+            $action_value === 'ALL_TODO_LIST' ||
+            $action_value === 'SELECT_TODO_LIST_TO_CHECK'
+        ) {
+            $title = '「' . $line_user->question->project->name . '」のやること一覧';
+            $text =   'プロジェクト:「' . $line_user->question->project->name . '」のやることは' . count($line_user->todo) . '件です';
+        } else if (
+            $action_value === 'WEEKLY_TODO_LIST' ||
+            $action_value === 'CHECK_TODO_BY_THIS_WEEK'
+        ) {
+            $title = $line_user->name . 'さんの今週までにやること一覧';
+            $text = $line_user->name . 'さんの今週までにやることは' . count($todo_carousel_columns) . '件です';
+        } else if ($action_value === 'CHECK_TODO_BY_TODAY') {
+            $title = $line_user->name . 'さんの今日までにやること一覧';
+            $text = $line_user->name . 'さんの今日までにやることは' . count($todo_carousel_columns) . '件です';
+        }
+        return ['title' => $title, 'text' => $text];
+    }
+
+    /**
+     * TodoListに戻る
+     *
+     * @param object $project
+     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
+     */
+    public static function comeBackTodoList(object $project)
+    {
+        $carouselText =  '「' . $project->name . '」のやること一覧に戻りますか？';
+        $actions = [
+            new PostbackTemplateActionBuilder('戻る', 'action=TODO_LIST&todo_uuid=' . $project->uuid),
+        ];
+        $builder = new CarouselColumnTemplateBuilder(null, $carouselText, null, $actions);
+        return $builder;
+    }
+
+
+    /**
+     *
+     *
+     * やることの期限 Date
+     *
+     *
+     */
 
     /**
      * Todoの期限を聞く
@@ -168,140 +350,6 @@ class Todo extends Model
     }
 
     /**
-     * 変更後のTodoの名前を聞く
-     *
-     * @param Todo $todo
-     * @return string $reply_message
-     */
-    public static function askTodoReName(Todo $todo)
-    {
-        return 'やること:「' . $todo->name . '」の変更後の名前を教えてください！';
-    }
-
-    /**
-     * どのTodoたちを振り返るか尋ねる
-     *
-     * @return \LINE\LINEBot\MessageBuilder\TemplateMessageBuilder
-     */
-    public static function askWhichCheckTodo()
-    {
-        $builder =
-            new TemplateMessageBuilder(
-                '振り返り', // チャット一覧に表示される
-                new ButtonTemplateBuilder(
-                    'どちらのやることを振り返りますか？', // title
-                    '選択してください', // text
-                    null, // 画像url
-                    [
-                        new PostbackTemplateActionBuilder('今日までにやること', 'action=CHECK_TODO_BY_TODAY&project_uuid='),
-                        new PostbackTemplateActionBuilder('今週までにやること', 'action=CHECK_TODO_BY_THIS_WEEK&todo_uuid='),
-                        new PostbackTemplateActionBuilder('やること一覧から選択', 'action=SELECT_TODO_LIST_TO_CHECK&todo_uuid='),
-                    ]
-                )
-            );
-        return $builder;
-    }
-
-    /**
-     * どのTodoたちを振り返るか尋ねる
-     *
-     * @param Todo $todo
-     * @return \LINE\LINEBot\MessageBuilder\TemplateBuilder\ConfirmTemplateBuilder
-     */
-    public static function askIfTodoHasBeenAccomplished(Todo $todo)
-    {
-        $text = '「' . $todo->name . '」について達成できましたか？';
-        $builder =
-            new ConfirmTemplateBuilder(
-                $text,
-                [
-                    new PostbackTemplateActionBuilder('はい', 'action=ACOMMPLISHED_TODO&todo_uuid=' . $todo->uuid),
-                    new PostbackTemplateActionBuilder('いいえ', 'action=NOT_ACCOMPLISHED_TODO&todo_uuid=' . $todo->uuid)
-                ]
-            );
-        return $builder;
-    }
-
-    /**
-     * 振り返りを続けるかどうか
-     *
-     * @param LineUsersQuestion $question
-     * @return \LINE\LINEBot\MessageBuilder\TemplateBuilder\ConfirmTemplateBuilder
-     */
-    public static function askContinueCheckTodo(LineUsersQuestion $question)
-    {
-        if ($question->checked_todo === CheckedTodo::CHECK_TODO['CHECK_TODO_BY_TODAY']) {
-            $title = '今日までにやることの振り返りを続けますか？';
-            $action_type = 'CHECK_TODO_BY_TODAY';
-        } else if ($question->checked_todo === CheckedTodo::CHECK_TODO['CHECK_TODO_BY_THIS_WEEK']) {
-            $title = '今週までにやることの振り返りを続けますか？';
-            $action_type = 'CHECK_TODO_BY_THIS_WEEK';
-        } else if ($question->checked_todo === CheckedTodo::CHECK_TODO['SELECT_TODO_LIST_TO_CHECK']) {
-            $title = '振り返りを続けますか？';
-            $action_type = 'SELECT_TODO_LIST_TO_CHECK';
-        }
-        $builder =
-            new ConfirmTemplateBuilder(
-                $title, // text
-                [
-                    new PostbackTemplateActionBuilder('続ける', 'action=' . $action_type . '&todo_uuid='),
-                    new PostbackTemplateActionBuilder('終了する', 'action=FINISH_CHECK_TODO&todo_uuid='),
-                ]
-            );
-        return $builder;
-    }
-
-    /**
-     * 振り返りを続けるかどうか
-     *
-     * @param LineUsersQuestion $question
-     * @return string $text
-     */
-    public static function getTextMessageOfFinishCheckTodo(LineUsersQuestion $question)
-    {
-        if ($question->checked_todo === CheckedTodo::CHECK_TODO['CHECK_TODO_BY_TODAY']) {
-            $text = '今日までにやることの振り返りを終了しました。';
-        } else if ($question->checked_todo === CheckedTodo::CHECK_TODO['CHECK_TODO_BY_THIS_WEEK']) {
-            $text = '今週までにやることの振り返りを終了しました。';
-        } else if ($question->checked_todo === CheckedTodo::CHECK_TODO['SELECT_TODO_LIST_TO_CHECK']) {
-            $text = '振り返りを終了しました。';
-        }
-        return $text;
-    }
-
-    /**
-     * どのTodoたちを振り返るか尋ねる
-     *
-     * @param Todo $todo
-     * @return \LINE\LINEBot\MessageBuilder\TemplateBuilder\ConfirmTemplateBuilder
-     */
-    public static function addTodoAfterCheckTodo(Todo $todo)
-    {
-        $text = '「' . $todo->name . '」を達成するためにやることを新しく追加しますか?';
-        $builder =
-            new ConfirmTemplateBuilder(
-                $text,
-                [
-                    new PostbackTemplateActionBuilder('はい', 'action=ADD_TODO_AFTER_CHECK_TODO&todo_uuid=' . $todo->uuid),
-                    new PostbackTemplateActionBuilder('いいえ', 'action=NOT_ADD_TODO_AFTER_CHECK_TODO&todo_uuid=' . $todo->uuid)
-                ]
-            );
-        return $builder;
-    }
-
-    /**
-     * 名前の変更を確認する
-     *
-     * @param Todo $todo
-     * @param DateTime $date
-     * @return string $reply_message
-     */
-    public static function reportNewTodoName(Todo $todo, string $new_todo_name)
-    {
-        return '「' . $todo->name . '」から「' . $new_todo_name . '」へ変更が完了しました';
-    }
-
-    /**
      * 日付を確認する
      *
      * @param Todo $todo
@@ -349,175 +397,5 @@ class Todo extends Model
                 )
             );
         return $builder;
-    }
-
-
-    /**
-     * Todo追加の呼びかけ
-     *
-     * @return string $reply_message
-     */
-    public static function callForAdditionalTodo()
-    {
-        return  'ゴールを達成するためにやることをどんどん追加していきましょう！' . "\n" . '「メニュー>やること」からゴール達成のためのやることを追加していくことができます！';
-    }
-
-    /**
-     * 振り返りの設定の説明
-     *
-     * @return string $reply_message
-     */
-    public static function explainSettingOfCheck()
-    {
-        return  'また週に1回、もくサポくんから振り返りのリマインドを送ることができるよ！' . "\n" . '「メニュー>振り返り」から日時の設定を自分で変更することができます！';
-    }
-
-    /**
-     * Todoの一覧表示
-     *
-     * @param User $line_user
-     * @param string $action_value
-     * @param array $todo_carousel_columns
-     * @return \LINE\LINEBot\MessageBuilder\TemplateMessageBuilder
-     */
-    public static function createTodoListTitleMessage(User $line_user, string $action_value, array $todo_carousel_columns)
-    {
-        if (
-            $action_value === 'ALL_TODO_LIST' ||
-            $action_value === 'SELECT_TODO_LIST_TO_CHECK'
-        ) {
-            $title = '「' . $line_user->question->project->name . '」のやること一覧';
-            $text =   'プロジェクト:「' . $line_user->question->project->name . '」のやることは' . count($line_user->todo) . '件です';
-        } else if (
-            $action_value === 'WEEKLY_TODO_LIST' ||
-            $action_value === 'CHECK_TODO_BY_THIS_WEEK'
-        ) {
-            $title = $line_user->name . 'さんの今週までにやること一覧';
-            $text = $line_user->name . 'さんの今週までにやることは' . count($todo_carousel_columns) . '件です';
-        } else if ($action_value === 'CHECK_TODO_BY_TODAY') {
-            $title = $line_user->name . 'さんの今日までにやること一覧';
-            $text = $line_user->name . 'さんの今日までにやることは' . count($todo_carousel_columns) . '件です';
-        }
-        return ['title' => $title, 'text' => $text];
-    }
-
-    /**
-     * Todoのカルーセル
-     *
-     * @param object $todo
-     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
-     */
-    public static function createTodoCarouselColumn(object $todo)
-    {
-        if ($todo->depth === "0") {
-            $parent = 'プロジェクト:「' . $todo->project->name . '」のゴール';
-        } else {
-            $parentTodo = Todo::where('uuid', $todo->parent_uuid)->first();
-            $parent = '「' . $parentTodo->name . '」のためにやること';
-        }
-        $accomplish = count($todo->accomplish) > 0 ? '【達成】' : '【未達成】';
-        $title = $accomplish . $todo->name;
-        $actions = [
-            new PostbackTemplateActionBuilder('名前・期限の変更/削除', 'action=CHANGE_TODO&todo_uuid=' . $todo->uuid),
-            new PostbackTemplateActionBuilder('やることの追加', 'action=ADD_TODO&todo_uuid=' . $todo->uuid),
-            new PostbackTemplateActionBuilder('振り返る', 'action=SELECT_CHECK_TODO&todo_uuid=' . $todo->uuid),
-        ];
-        $builder = new CarouselColumnTemplateBuilder($title, $parent, null, $actions);
-        return $builder;
-    }
-
-    /**
-     * CheckTodoのカルーセル
-     *
-     * @param object $todo
-     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
-     */
-    public static function createCheckTodoCarouselColumn(object $todo)
-    {
-        if ($todo->depth === "0") {
-            $parent = 'プロジェクト:「' . $todo->project->name . '」のゴール';
-        } else {
-            $parentTodo = Todo::where('uuid', $todo->parent_uuid)->first();
-            $parent = '「' . $parentTodo->name . '」のためにやること';
-        }
-        $accomplish = count($todo->accomplish) > 0  ? '【達成】' : '【未達成】';
-        $title = $accomplish . $todo->name;
-        $actions = [
-            new PostbackTemplateActionBuilder('振り返る', 'action=CHECK_TODO&todo_uuid=' . $todo->uuid),
-        ];
-        $builder = new CarouselColumnTemplateBuilder($title, $parent, null, $actions);
-        return $builder;
-    }
-
-    /**
-     * 作ったTodoのTodoを新しく追加する
-     *
-     * @param object $todo
-     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
-     */
-    public static function continueAddTodoOfTodo(object $todo)
-    {
-        $carouselText =  '「' . $todo->name . '」を達成するためにやることを新しく追加しますか?';
-        $actions = [
-            new PostbackTemplateActionBuilder('追加する', 'action=ADD_TODO&todo_uuid=' . $todo->uuid),
-        ];
-        $builder = new CarouselColumnTemplateBuilder(null, $carouselText, null, $actions);
-        return $builder;
-    }
-
-    /**
-     * 引き続き親TodoのTodoを追加する
-     *
-     * @param object $parentTodo
-     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
-     */
-    public static function continueAddTodoOfParentTodo(object $parent_todo)
-    {
-        $carouselText =  '「' . $parent_todo->name . '」を達成するためにやることを引き続き追加しますか?';
-        $actions = [
-            new PostbackTemplateActionBuilder('追加する', 'action=ADD_TODO&todo_uuid=' . $parent_todo->uuid),
-        ];
-        $builder = new CarouselColumnTemplateBuilder(null, $carouselText, null, $actions);
-        return $builder;
-    }
-
-    /**
-     * TodoListに戻る
-     *
-     * @param object $project
-     * @return LINE\LINEBot\MessageBuilder\TemplateBuilder\CarouselColumnTemplateBuilder;
-     */
-    public static function comeBackTodoList(object $project)
-    {
-        $carouselText =  '「' . $project->name . '」のやること一覧に戻りますか？';
-        $actions = [
-            new PostbackTemplateActionBuilder('戻る', 'action=TODO_LIST&todo_uuid=' . $project->uuid),
-        ];
-        $builder = new CarouselColumnTemplateBuilder(null, $carouselText, null, $actions);
-        return $builder;
-    }
-
-    /**
-     * Todoを変更する
-     *
-     * @param Todo $todo
-     * @return \LINE\LINEBot\MessageBuilder\TemplateMessageBuilder
-     */
-    public static function changeTodo(Todo $todo)
-    {
-        $title =  '「' . $todo->name . '」';
-        return new TemplateMessageBuilder(
-            $title,
-            new ButtonTemplateBuilder(
-                $title,
-                "選択してください",
-                null,
-                [
-                    new PostbackTemplateActionBuilder("名前の変更", 'action=RENAME_TODO&todo_uuid=' . $todo->uuid),
-                    new PostbackTemplateActionBuilder('やることの削除', 'action=DELETE_TODO&todo_uuid=' . $todo->uuid),
-                    new PostbackTemplateActionBuilder('期限の変更', 'action=ASK_RESCHEDULE&todo_uuid=' . $todo->uuid),
-                ]
-            )
-        );
     }
 }
